[Player] Improve seeking
* Use yt-dlp --download-sections flag to seek instead of seeking after downloading
* Create getSecondsFromUrlTimestamp util to parse URL timestamp from Twitch or YouTube into a raw number of seconds
* Support URL timestamps for tracks to seek
* Rename duration fields in currentTrackPlayTime to indicate the unit of measurement (milliseconds)
  for better documentation and to reduce future bugs

Seeking with yt-dlp means that the yt-dlp process
will only download the sections of the stream needed.
Seeking with FFmpeg ourselves means that the entire stream must be downloaded first and then seeked,
which is both wasteful and time consuming.
yt-dlp will do the seeking itself (with FFmpeg) while downloading.