Feat(Movies): Add NOT UUID validation to custom_id col
This is so that the custom_id cannot be set to
the id of another column. Doing this at the
application level instead of the database
level is fine. It's not a big concern and is not
worth the effort or performance to properly
validate this at the database level.